refactor: visualizer element
